requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0

# LangChain Core
langchain>=0.1.0
langchain-community>=0.0.13
langchain-google-genai>=0.0.6

# Embeddings
langchain-huggingface>=0.0.1
sentence-transformers>=2.3.1

# Vector Database
faiss-cpu

# PDF Support
pypdf>=3.17.4

# Utilities
# Relaxing pydantic to allow Python 3.13 compatible versions
pydantic>=2.7.0
python-multipart>=0.0.6
aiofiles>=23.2.1